refactor: ConsentsNewsletters to split out state
